Refactor `RoxButton` into separate components `RowButtons` and `RowButton`.
DDFFORM-3
