+++ base.scss
@@ -98,6 +98,7 @@
 @import "./src/stories/Library/button-box/button-box";
 @import "./src/stories/Library/links/arrow-link/arrow-link";
 @import "./src/stories/Library/Buttons/row-button/row-button";
+@import "./src/stories/Library/Buttons/row-button/row-buttons";
 @import "./src/stories/Library/article-header/article-header";
 
 // Autosuggest block styling needs to be loaded before the rest of the scss for autosuggest
