+++ src/stories/Library/Buttons/row-button/RowButton.tsx
@@ -1,30 +1,16 @@
-import clsx from "clsx";
 import { FC } from "react";
 
 type RowButtonProps = {
-  labels: string[];
-  className?: string;
+  label: string;
 };
 
-const RowButton: FC<RowButtonProps> = ({ labels, className }) => (
-  <div className={clsx("row-buttons", className)}>
-    {labels.slice(0, 2).map((label) => (
-      <button
-        className="row-button text-tags row-button__text capitalize-all"
-        type="button"
-      >
-        {label}
-      </button>
-    ))}
-    {labels.length > 2 && (
-      <button
-        className="row-button text-tags row-button__text capitalize-all"
-        type="button"
-      >
-        ...
-      </button>
-    )}
-  </div>
+const RowButton: FC<RowButtonProps> = ({ label }) => (
+  <button
+    className="row-button text-tags row-button__text capitalize-all"
+    type="button"
+  >
+    {label}
+  </button>
 );
 
 export default RowButton;
