+++ src/stories/Library/Buttons/row-button/row-button.scss
@@ -1,8 +1,3 @@
-.row-buttons {
-  display: flex;
-  flex-direction: row;
-}
-
 .row-button {
   height: 40px;
   background-color: $c-global-secondary;
