+++ src/stories/Library/Buttons/row-button/row-buttons.scss
@@ -1,0 +1,4 @@
+.row-buttons {
+  display: flex;
+  flex-direction: row;
+}
